supp .env, fin de la facto et debut users session
