--- server-sora/.env
@@ -1,4 +1,0 @@
-MONGO_URI=mongodb+srv://burak:burakbecode@cluster0.8gjkupo.mongodb.net/soracine
-
-
-MOVIE_DB_API_KEY=c8e37a819c856aff51d6efd1114d3534
