+++ server-sora/.env.exemple
@@ -1,0 +1,5 @@
+#MONGO_URI
+
+#API_KEY
+
+#TOKEN_SECRET
