+++ server-sora/.gitignore
@@ -1,4 +1,10 @@
-# .gitignore
+# dependencies
 node_modules/
+
+
+# environment variables
 .env
 config.js
+
+
+
