+++ server-sora/controllers/connectDB.js
@@ -1,0 +1,26 @@
+const mongoose = require("mongoose");
+
+function connectDB() {
+  mongoose.connect(process.env.MONGO_URI, {
+    useNewUrlParser: true,
+    useUnifiedTopology: true,
+  });
+  const db = mongoose.connection;
+  db.on("error", err => {
+    console.error("MongoDB connection error:", err);
+  });
+  db.once("open", () => {
+    console.log("MongoDB connected successfully");
+  });
+  db.on("disconnected", () => {
+    console.log("MongoDB disconnected");
+  });
+  process.on("SIGINT", () => {
+    db.close(() => {
+      console.log("MongoDB connection closed");
+      process.exit(0);
+    });
+  });
+}
+
+module.exports = connectDB;
