+++ server-sora/package.json
@@ -20,6 +20,7 @@
     "express": "^4.18.2",
     "express-session": "^1.17.3",
     "fs": "0.0.1-security",
+    "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.0.2",
     "nodemailer": "^6.9.7",
     "nodemon": "^3.0.1"
